renamed sync ts file
