+++ src/core/virtual-circuit/components/rfid.sync.ts
@@ -1,4 +1,4 @@
-import { SyncComponent, ResetComponent } from '../svg.component';
+import { SyncComponent, ResetComponent } from '../svg-sync';
 import { CreateComponentHook, CreateWire } from '../svg-create';
 
 import { RfidState } from '../../frames/arduino-components.state';
