+++ src/core/virtual-circuit/svg-sync.ts
